fixed score multiplier
